doubles by default, ! bug in float version

+++ include/Stk.h
@@ -81,7 +81,7 @@ namespace stk {
 // following user-definable floating-point type.  You
 // can change this to "float" if you prefer or perhaps
 // a "long double" in the future.
-typedef float StkFloat;
+typedef double StkFloat;
 
 //! STK error handling class.
 /*!
